詳細と関連動画をPromise.all
