+++ src/components/Header/Header.module.scss
@@ -6,6 +6,7 @@
   height: 56px;
   width: 100%;
   background-color: #202020;
+  z-index: 99;
   a{
     color: #fff;
     font-size: 1.2rem;
