+++ src/components/Layout/Layout.module.scss
@@ -1,5 +1,4 @@
 .wrapper{
-  min-height: 100vh;
   display: flex;
   flex-direction: column;
   background: #181818;
@@ -10,4 +9,9 @@
   flex:1;
   width: 100%;
   background: #181818;
+  @media screen and (max-width:  960px) {
+    min-width: 90%;
+    margin: 0 auto;
+    flex-direction: column;
+  }
 }
