+++ src/components/SideList/SideList.js
@@ -1,25 +1,10 @@
-import React, {useEffect, useContext} from 'react'
+import React, {useContext} from 'react'
 import { Store } from '../../store/index'
-import { fetchRelatedData } from '../../apis/index'
 import SideListItem from '../SideListItem/SideListItem';
 import Style from "./SideList.module.scss"
 
 const SideList = () => {
-  const {globalState, setGlobalState} = useContext(Store);
-  const setRelatedVideo = async (id) =>{
-    await fetchRelatedData(id).then((res)=>{
-      setGlobalState({
-        type: "SET_RELATED",
-        payload: {
-          related: res.data.items
-        }
-      })
-    })
-  }
-  useEffect(()=>{
-    setRelatedVideo(globalState.selected.id);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  },[globalState.selected])
+  const {globalState} = useContext(Store);
   return (
     <div className={Style.sidenav}>
       { globalState.related ? globalState.related.map((video)=>{
