+++ src/components/VideoDetail/VideoDetail.js
@@ -1,34 +1,11 @@
-import React, { useContext, useEffect } from 'react'
-// 現在のURLのpathやサーチパラメータにアクセスしやすくなる
-import { useLocation } from 'react-router-dom'
-import { fetchSelectedData } from '../../apis/index';
+import React, { useContext } from 'react'
 import { Store } from '../../store';
 import VideoPlay from '../VideoPlay/VideoPlay';
 import Linkify from 'react-linkify'
 import Style from "./VideoDetail.module.scss"
 
 const VideoDetail = () => {
-  const {globalState, setGlobalState} = useContext(Store);
-  const location = useLocation();
-  const setSelectedVideo = async () =>{
-    const searchParams = new URLSearchParams(location.search)
-    const id = searchParams.get('v');
-    await fetchSelectedData(id).then((res)=>{
-      const item = res.data.items.shift();
-      setGlobalState({
-        type: "SET_SELECTED",
-        payload: {
-          selected: item
-        }
-      });
-      console.log(item);
-    });
-  }
-
-  useEffect(()=>{
-    setSelectedVideo();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  },[location])
+  const {globalState} = useContext(Store);
   return ( globalState.selected && globalState.selected.id ? (
     <div className={Style.wrap}>
       <VideoPlay id={globalState.selected.id} />
